+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.7] - 2025-05-28
+## [v1.2.8] - 2025-06-06
+
+### Fixed
+
+- Removing obsolete filters from `filters.json` and `filters_i18n.json` metadata
+  for `mac_v2` platform.
+
+[v1.2.8]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.2.7...v1.2.8
+
+## [v1.2.7] - 2025-05-28
 
 ### Added